test_process_replies.py
import process_replies

def test_apply_diff():
    # Test case for applying diff
    original_file = 'test_file.txt'
    with open(original_file, 'w') as f:
        f.write('This is a test file.\n')
    diff = '--- test_file.txt\n+++ test_file.txt\n@@ -1,1 +1,2 @@\n This is a test file.\n+This is a new line.\n'
    process_replies.apply_diff(diff)
    with open(original_file, 'r') as f:
        content = f.read()
    assert content == 'This is a test file.\nThis is a new line.\n'

def test_process_diff():
    # Test case for processing diff command
    reply_data = {
        'command': 'diff',
        'contents': '--- test_file.txt\n+++ test_file.txt\n@@ -1,2 +1,3 @@\n This is a test file.\n This is a new line.\n+Another new line.\n',
        'reasoning': 'Adding another new line to the test file.',
        'id': 3
    }
    result = process_replies.process_diff(reply_data['contents'], reply_data['reasoning'], reply_data['id'])
    assert result == reply_data

def test_process_file_write():
    # Test case for processing file_write command
    reply_data = {
        'command': 'file_write',
        'contents': {'new_file.txt': 'This is a new file.'},
        'reasoning': 'Creating a new file for testing purposes.',
        'id': 4
    }
    result = process_replies.process_file_write(reply_data['contents'], reply_data['reasoning'], reply_data['id'])
    assert result == reply_data

def test_process_shell():
    # Test case for processing shell command
    reply_data = {
        'command': 'shell',
        'contents': 'echo "Hello, World!"',
        'reasoning': 'Testing shell command execution.',
        'id': 5
    }
    result = process_replies.process_shell(reply_data['contents'], reply_data['reasoning'], reply_data['id'])
    assert result['output'] == 'Hello, World!\n'

def test_process_reply():
    # Test case for processing reply
    reply_data = {
        'command': 'shell',
        'contents': 'echo "Hello, World!"',
        'reasoning': 'Testing shell command execution.',
        'id': 6
    }
    result = process_replies.process_reply(reply_data)
    assert result['output'] == 'Hello, World!\n'